+++ src/main/java/com/mulmul/todo/post/dto/bundle/PostDeleteBundle.java
@@ -1,4 +1,4 @@
-package com.mulmul.todo.dto.bundle;
+package com.mulmul.todo.post.dto.bundle;
 
 public class PostDeleteBundle {
     private final Long id;
