+++ src/main/java/com/mulmul/todo/post/dto/bundle/PostUpdateBundle.java
@@ -1,0 +1,25 @@
+package com.mulmul.todo.post.dto.bundle;
+
+public class PostUpdateBundle {
+    private final Long id;
+    private final String title;
+    private final String content;
+
+    public PostUpdateBundle(Long id, String title, String content) {
+        this.id = id;
+        this.title = title;
+        this.content = content;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public String getContent() {
+        return content;
+    }
+}
