+++ src/main/java/com/mulmul/todo/post/dto/response/PostCreateResponse.java
@@ -1,17 +1,15 @@
-package com.mulmul.todo.dto.response;
+package com.mulmul.todo.post.dto.response;
 
-import com.mulmul.todo.domain.vo.Status;
+import com.mulmul.todo.post.domain.vo.Status;
 
-public class PostDetailResponse {
+public class PostCreateResponse {
     private final Long id;
     private final String title;
-    private final String content;
     private final String status;
 
-    public PostDetailResponse(Long id, String title, String content, Status status) {
+    public PostCreateResponse(Long id, String title, Status status) {
         this.id = id;
         this.title = title;
-        this.content = content;
         this.status = status.name();
     }
 
@@ -21,10 +19,6 @@ public class PostDetailResponse {
 
     public String getTitle() {
         return title;
-    }
-
-    public String getContent() {
-        return content;
     }
 
     public String getStatus() {
