+++ src/main/java/com/mulmul/todo/post/dto/response/PostDeleteResponse.java
@@ -1,4 +1,4 @@
-package com.mulmul.todo.dto.response;
+package com.mulmul.todo.post.dto.response;
 
 public class PostDeleteResponse {
     private final Long id;
