+++ src/main/java/com/mulmul/todo/post/dto/response/PostDetailResponse.java
@@ -1,14 +1,18 @@
-package com.mulmul.todo.dto.bundle;
+package com.mulmul.todo.post.dto.response;
 
-public class PostUpdateBundle {
+import com.mulmul.todo.post.domain.vo.Status;
+
+public class PostDetailResponse {
     private final Long id;
     private final String title;
     private final String content;
+    private final String status;
 
-    public PostUpdateBundle(Long id, String title, String content) {
+    public PostDetailResponse(Long id, String title, String content, Status status) {
         this.id = id;
         this.title = title;
         this.content = content;
+        this.status = status.name();
     }
 
     public Long getId() {
@@ -21,5 +25,9 @@ public class PostUpdateBundle {
 
     public String getContent() {
         return content;
+    }
+
+    public String getStatus() {
+        return status;
     }
 }
